Moved utils and resources top level
