+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/base/app/BaseLoadingFragment.java
@@ -13,7 +13,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.base.utils.Result;
+import com.ninetwozero.bf4intel.utils.Result;
 
 public abstract class BaseLoadingFragment extends BaseFragment implements LoaderManager.LoaderCallbacks<Result> {
     private Gson gson = new Gson();
