+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/ui/battlereport/BattleReportAdapter.java
@@ -10,9 +10,9 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.base.resource.maps.GameModeStringMap;
-import com.ninetwozero.bf4intel.base.resource.maps.LevelStringMap;
-import com.ninetwozero.bf4intel.base.utils.DateTimeUtils;
+import com.ninetwozero.bf4intel.resource.maps.GameModeStringMap;
+import com.ninetwozero.bf4intel.resource.maps.LevelStringMap;
+import com.ninetwozero.bf4intel.utils.DateTimeUtils;
 import com.ninetwozero.bf4intel.json.battlereports.SummaryBattleReport;
 import com.ninetwozero.bf4intel.json.battlereports.Team;
