+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/ui/fragments/SoldierOverviewFragment.java
@@ -15,14 +15,14 @@ import com.ninetwozero.bf4intel.base.connection.ConnectionRequest;
 import com.ninetwozero.bf4intel.base.connection.IntelLoader;
 import com.ninetwozero.bf4intel.base.datatypes.Skill;
 import com.ninetwozero.bf4intel.base.factories.UrlFactory;
-import com.ninetwozero.bf4intel.base.resource.Keys;
-import com.ninetwozero.bf4intel.base.resource.maps.CompletionStringMap;
-import com.ninetwozero.bf4intel.base.resource.maps.RankStringMap;
-import com.ninetwozero.bf4intel.base.resource.maps.VehicleStringMap;
-import com.ninetwozero.bf4intel.base.resource.maps.WeaponStringMap;
-import com.ninetwozero.bf4intel.base.utils.DateTimeUtils;
-import com.ninetwozero.bf4intel.base.utils.PersonaUtils;
-import com.ninetwozero.bf4intel.base.utils.Result;
+import com.ninetwozero.bf4intel.resource.Keys;
+import com.ninetwozero.bf4intel.resource.maps.CompletionStringMap;
+import com.ninetwozero.bf4intel.resource.maps.RankStringMap;
+import com.ninetwozero.bf4intel.resource.maps.VehicleStringMap;
+import com.ninetwozero.bf4intel.resource.maps.WeaponStringMap;
+import com.ninetwozero.bf4intel.utils.DateTimeUtils;
+import com.ninetwozero.bf4intel.utils.PersonaUtils;
+import com.ninetwozero.bf4intel.utils.Result;
 import com.ninetwozero.bf4intel.json.soldieroverview.BaseStatsModel;
 import com.ninetwozero.bf4intel.json.soldieroverview.CompletionProgress;
 import com.ninetwozero.bf4intel.json.soldieroverview.SkillOverview;
